Changed parts type to a vector of integers. Added Maya calendars (not fully working yet!)

+++ plugins/Calendars/src/GregorianCalendar.hpp
@@ -39,7 +39,7 @@ public:
 
 	//! set date from a vector of calendar date elements sorted from the largest to the smallest.
 	//! Year-Month[1...12]-Day[1...31]-Hour-Minute-Second
-	virtual void setParts(QVector<double> parts) Q_DECL_OVERRIDE;
+	virtual void setParts(QVector<int> parts) Q_DECL_OVERRIDE;
 
 	//! returns true for leap years
 	static bool isLeap(int year);
